Core/Auras: Implemented using all aura interrupt flag fields

+++ Source/Game/Combat/ThreatManager.cs
@@ -381,7 +381,7 @@ namespace Game.Combat
                 Contract.Assert(target);                                     // if the ref has status online the target must be there !
 
                 // some units are prefered in comparison to others
-                if (!noPriorityTargetFound && (target.IsImmunedToDamage(attacker.GetMeleeDamageSchoolMask()) || target.HasNegativeAuraWithInterruptFlag((uint)SpellAuraInterruptFlags.TakeDamage)))
+                if (!noPriorityTargetFound && (target.IsImmunedToDamage(attacker.GetMeleeDamageSchoolMask()) || target.HasNegativeAuraWithInterruptFlag(SpellAuraInterruptFlags.TakeDamage)))
                 {
                     if (i != threatList.Count - 1)
                     {
